Modify exam_result_tables to suit changes in Excel form

+++ ft/exam_result_tables.py
@@ -2,6 +2,7 @@ import os
 import csv
 import shutil
 import argparse
+from typing import List, Union
 
 from .formats import Table
 
@@ -14,7 +15,14 @@ def main():
     parser.add_argument(
         '-s', '--short-format',
         action='store_true',
+        default=False,
         help='Include only the total score, not the score on individual tasks.'
+    )
+    parser.add_argument(
+        '--keep-tex',
+        action='store_true',
+        default=False,
+        help='Keep tex files after compilation.'
     )
     parser.add_argument(
         '--header',
@@ -37,12 +45,16 @@ def main():
         print(f'Creating directory {temp_tex_source_dir}')
         os.mkdir(temp_tex_source_dir)
 
-    result_table = read_source_csv(in_fname, args.short_format)
+    result_table = prepare_raw_table(
+        read_source_csv(in_fname),
+        args.short_format
+    )
     create_student_tex_files(temp_tex_source_dir, result_table)
     create_main_tex_and_compile(temp_tex_source_dir, out_fname, args.header)
 
-    print(f'Removing directory {temp_tex_source_dir}')
-    shutil.rmtree(temp_tex_source_dir)
+    if not args.keep_tex:
+        print(f'Removing directory {temp_tex_source_dir}')
+        shutil.rmtree(temp_tex_source_dir)
 
 
 def create_output_fname(in_fname: str):
@@ -50,42 +62,62 @@ def create_output_fname(in_fname: str):
     return f_name + '.pdf'
 
 
-def read_source_csv(f_name: str, short_format: bool = False):
+def columns_to_keep(table: List[List[str]], short_format: bool) -> List[int]:
+    num_cols = len(table[0])
+    # name and task scores, stop when totals are passed
+    result = [i for i in range(num_cols - 7) if table[1][i] != '']
+    if short_format:
+        result = [0] + result[-6:]  # keep only name and total scores
+    result += [
+        num_cols - 6,   # grade
+        num_cols - 1    # comment for student
+    ]
+    return result
+
+
+def read_source_csv(f_name: str) -> List[List[str]]:
     if not os.path.isfile(f_name):
         raise FileNotFoundError(f'{f_name} is not a file')
 
-    with open(f_name, 'r') as infile:
+    with open(f_name, 'r', encoding='UTF-8') as infile:
         reader = csv.reader(infile, delimiter='\t')
-        table = [line for line in reader if line[-4] != '']
-    cols_to_keep = [i for i in range(len(table[0]) - 2) if table[1][i] != ''] \
-        + [len(table[0]) - 1]
-    if short_format:
-        cols_to_keep = [0] + cols_to_keep[-7:]
-    table = [[table[i][j].replace('%', r'\%') for j in cols_to_keep]
-             for i in range(len(table))]
-    table[1][0] = 'Oppgave'
-    table[2][0] = 'Mulige poeng'
-    return table
+        return [line for line in reader if line[-9] != '']
+
+
+def prepare_raw_table(
+    table: List[List[str]],
+    short_format: bool
+) -> Union[List[List[str]], str]:
+    cols = columns_to_keep(table, short_format)
+    return [[row[i].replace('%', r'\%') for i in cols] for row in table]
 
 
 def create_student_tex_files(temp_tex_dir: str, table):
+    header_rows = [row[:-1] for row in table[:3]]   # drop comment field
+    # remove percentages for totals scores on part 1 and 2
+    header_rows[2][-4] = ''
+    header_rows[2][-6] = ''
+
     fnames = []
-    header_rows = table[:3]
     for student in table[3:]:
         student_name = student[0]
-        student_table = header_rows + [student]
+        student_table = header_rows + [student[:-1]]
         tab = Table(student_table)
         tab.transpose()
-        fname = student_name.replace(' ', '_')
-        fname = f'{fname}.tex'
+        fname = student_name.replace(' ', '_') + '.tex'
         fnames.append(fname)
+        tex_fname = os.path.join(temp_tex_dir, fname)
         num_rows = len(header_rows[0])
         tab.latex(
-            os.path.join(temp_tex_dir, fname),
+            tex_fname,
             extra_separator=[0, num_rows - 8, num_rows - 6, num_rows - 4]
         )
+        comment = student[-1]
+        if comment:
+            with open(tex_fname, 'a', encoding='UTF-8') as f:
+                f.write('\n' + r'\vspace{1cm}' + r'\textbf{Kommentar:} ' + comment)
 
-    with open(os.path.join(temp_tex_dir, 'student_list.tex'), 'w') as outfile:
+    with open(os.path.join(temp_tex_dir, 'student_list.tex'), 'w', encoding='UTF-8') as outfile:
         s = '\n' + r'\newpage' + '\n'
         outfile.write(s.join([r'\input{' + fname + '}' for fname in fnames]))
